Extract TomlTree conversion to its own file
